+++ docs/CHANGE_LOG.md
@@ -2,6 +2,7 @@
 
 ## Version 2.3.2:
 - When pressing escape, if events are selected and another view is opened, the events selected are cleared as the first action (the view will remain open).
+- When pressing escape, if the "Search" dialogs search history drop down is open, it is closed (instead of closing the dialog itself).
 - 
 
 <br>